feat(testbed): F52 browser test harness -- wasm-bindgen in the testbed
- Add wasm-bindgen-test feature to foundation_testbed: re-exports js-sys,
  web-sys, wasm-bindgen-futures so consumer crates add one dev-dep
- New wasm-testbed bindgen command: build, wasm-bindgen --target web,
  serve, Chromium CDP, poll output
- Wasm binary discovery: scan ALL .wasm in deps/ for __wbgt_ exports
- Fix wasm-bindgen --target web flag (was defaulting to bundler, no init)
- Browser driver diagnostic on timeout (title, hasOutput, bodyText)
- Split foundation_netio dev-deps: native-only, wasm32-only with bindgen
- WebSocket browser tests in foundation_netio/tests/wasm/
- Fix wasm32 overflow in timeout.rs (u64-then-cast)
- Getrandom js feature for wasm32 (transitive via walrus)
- Make which optional (not wasm32-compatible)
- CLI rename: Web to Browser, new Bindgen command
- Version check: CLI >= crate version, single-source-of-truth pins
- Docs: all three test paths (owned, bindgen, native Harness), README
- Feature spec: specifications/41-connectrpc/features/52-browser-test-harness

Verified: 170 native tests pass, 2 browser tests in Chromium pass,
wasm32 compiles zero errors.

+++ backends/foundation_netio/src/shared/http/timeout.rs
@@ -735,7 +735,7 @@ mod tests {
 
         // Huge body (36GB+) gets max timeout (60s)
         // sqrt(36M KB) * 10ms = 6000 * 10ms = 60s, clamped to max
-        let ctx = TimeoutContext::with_size(36 * 1024 * 1024 * 1024);
+        let ctx = TimeoutContext::with_size((36u64 * 1024 * 1024 * 1024) as usize);
         let timeout = calc.calculate_read_timeout(&ctx);
         assert_eq!(timeout, Duration::from_secs(60));
     }
@@ -776,7 +776,7 @@ mod tests {
         // Huge body (75GB+) gets clamped to max_read_timeout (60s)
         // sqrt(75M KB) * 10ms = 8660 * 10ms = 86.6s per read, clamped to 60s
         // With 4 attempts: 60s * 4 = 240s
-        let ctx = TimeoutContext::with_size(75 * 1024 * 1024 * 1024);
+        let ctx = TimeoutContext::with_size((75u64 * 1024 * 1024 * 1024) as usize);
         let timeout = calc.calculate_total_timeout(&ctx);
         assert_eq!(timeout, Duration::from_secs(240));
     }
@@ -797,7 +797,7 @@ mod tests {
 
         // Huge body (9GB+) uses custom max (30s)
         // sqrt(9M KB) * 10ms = 3000 * 10ms = 30s, clamped to max
-        let ctx = TimeoutContext::with_size(9 * 1024 * 1024 * 1024);
+        let ctx = TimeoutContext::with_size((9u64 * 1024 * 1024 * 1024) as usize);
         let timeout = calc.calculate_read_timeout(&ctx);
         assert_eq!(timeout, Duration::from_secs(30));
     }
